This is a minor error with it leaving two voice channels instead of one. for now it will do.

+++ Dikubot/Discord/EventListeners/ExpandableVoiceChatListener.cs
@@ -11,7 +11,7 @@ namespace Dikubot.Discord.EventListeners
     public class ExpandableVoiceChatListener
     {
 	    /// <Summary>Will delete all the empty channels.</Summary>
-	    /// <param name="channel">The channels group that will be deleted from.</param>
+	    /// <param name="channel">The channel that will be deleted from.</param>
 	    /// <return>Task.</return>
 	    private async Task DeleteAllEmptyExpandMembers(SocketVoiceChannel channel)
 	    {
@@ -27,21 +27,53 @@ namespace Dikubot.Discord.EventListeners
 			    return;
 
 		    var children = voiceChannelServices.GetAll(m => m.DeleteOnLeave && m.ExpandId == channelModel.ExpandId);
-		    var voiceChannels = new List<SocketVoiceChannel>();
 
-		    foreach (var child in children)
+		    var childrenArray = children.ToArray();
+		    for (int i = 1; i < childrenArray.Length; i++)
 		    {
-			    var gottenChannel = guild.GetVoiceChannel(Convert.ToUInt64(child.DiscordId));
-			    if (gottenChannel?.Users?.Count != null && gottenChannel.Users.Count == 0)
-				    voiceChannels.Add(gottenChannel);
-		    }
-
-		    foreach (var voiceChannel in voiceChannels)
-		    {
-			    await permissionsService.RemoveVoiceChannel(voiceChannel);
+			    var voiceChannel = guild.GetVoiceChannel(Convert.ToUInt64(childrenArray[i].DiscordId));
+			    if (voiceChannel?.Users?.Count != null && voiceChannel.Users.Count == 0)
+				    await permissionsService.RemoveVoiceChannel(voiceChannel);
 		    }
 	    }
     
+	    /// <Summary>Create a channel if no channel is availbe.</Summary>
+	    /// <param name="channel">The channel that will be used to create from.</param>
+	    /// <return>Task.</return>
+	    private async Task AddChannelIfNonIsAvaiable(SocketVoiceChannel channel)
+	    {
+		    var voiceChannelServices = new VoiceChannelServices();
+		    
+		    // If channel is null return.
+		    if (channel == null)
+			    return;
+		    var model = voiceChannelServices.Get(m => m.DiscordId == channel.Id.ToString());
+
+		    // If channel isn't expandable return.
+		    if (model?.ExpandOnJoin == null || !model.ExpandOnJoin)
+			    return;
+		    
+		    // If a single voice channel is empty return.
+		    var children = voiceChannelServices.GetAll(m => m.ExpandId == model.ExpandId);
+		    foreach (var child in children)
+		    {
+			    var gottenChannel = channel.Guild.GetVoiceChannel(Convert.ToUInt64(child.DiscordId));
+			    if (gottenChannel?.Users?.Count != null && gottenChannel.Users.Count == 0)
+				    return;
+		    }
+		    // Makes a new voice chanel model
+		    var childModel = new VoiceChannelModel();
+		    childModel.ExpandOnJoin = true;
+		    childModel.DeleteOnLeave = true;
+		    childModel.Position = model.Position;
+		    childModel.ExpandId = model.ExpandId;
+		    childModel.Name = "Rum";
+		    childModel.DiscordCategoryId = model.DiscordCategoryId;
+
+		    // Adds the voice channel to the database and server.
+		    var permissionsService = new PermissionsService(channel.Guild);
+		    await permissionsService.AddVoiceChannel(childModel);
+	    }
 	    
 	    /// <Summary>Makes a new voice chat when someone joins a expandable voice chat.</Summary>
 		/// <param name="user">The user who joined.</param>
@@ -52,41 +84,10 @@ namespace Dikubot.Discord.EventListeners
 	    {
 		    var leaveChannel = leaveState.VoiceChannel;
 		    var joinChannel = joinState.VoiceChannel;
+		    await AddChannelIfNonIsAvaiable(leaveChannel);
 		    await DeleteAllEmptyExpandMembers(leaveChannel);
-		    await DeleteAllEmptyExpandMembers(joinChannel);
-
-		    // If joinChannel is null return.
-		    if (joinChannel == null)
-			    return;
-
-		    var voiceChannelServices = new VoiceChannelServices();
-		    var model = voiceChannelServices.Get(m => m.DiscordId == joinChannel.Id.ToString());
+		    await AddChannelIfNonIsAvaiable(joinChannel);
 		    
-		    // If channel isn't expandable return.
-		    if (model?.ExpandOnJoin == null || !model.ExpandOnJoin)
-			    return;
-
-		    // if a single voice channel is empty return;
-		    var children = voiceChannelServices.GetAll(m => m.ExpandId == model.ExpandId);
-			foreach (var child in children)
-			{
-				var gottenChannel = joinChannel.Guild.GetVoiceChannel(Convert.ToUInt64(child.DiscordId));
-				if (gottenChannel?.Users?.Count != null && gottenChannel.Users.Count == 0)
-					return;
-			}
-
-			// Makes a new voice chanel model
-			var childModel = new VoiceChannelModel();
-			childModel.ExpandOnJoin = true;
-			childModel.DeleteOnLeave = true;
-			childModel.Position = model.Position;
-			childModel.ExpandId = model.ExpandId;
-			childModel.Name = "Rum";
-			childModel.DiscordCategoryId = model.DiscordCategoryId;
-
-			// Adds the voice channel to the database and server.
-			var permissionsService = new PermissionsService(joinChannel.Guild);
-			await permissionsService.AddVoiceChannel(childModel);
 	    }
     }
 }
